Explicitly request notification permissions and move firebase.ts
