+++ client/src/main.tsx
@@ -8,7 +8,7 @@ import { Provider } from "react-redux";
 
 import "./index.css";
 import App from "./app/App";
-import { configure_firebase } from "./firebase";
+import { configure_firebase } from "./app/firebase";
 
 configure_firebase();
